Chrome Debug Fails to Start #588
When Chrome/Firefox Debug launch is created, the only browser defined in 'Run with:` field of launch configuration
is to be used to start debuging. So, no any additional searches for a browser executable to be performed.
If no browser is confiured, a message is displayed prompting a user to open Launch Configuration Browser Tab or Web Browser
preference page to configure/select a correct browser.

Firefox Debug is also updated to serach fora browser executable in a list of browsers defined in Web Browser preference page.
If no executable can be found it is suggested to open Web Browser preference page to add/edit Firefox browser location.

Fixes: #588
Signed-off-by: Victor Rubezhny <vrubezhny@redhat.com>